+++ dev/kyuubi-codecov/pom.xml
@@ -46,6 +46,13 @@
 
         <dependency>
             <groupId>org.apache.kyuubi</groupId>
+            <artifactId>kyuubi-zookeeper</artifactId>
+            <version>${project.version}</version>
+            <scope>test</scope>
+        </dependency>
+
+        <dependency>
+            <groupId>org.apache.kyuubi</groupId>
             <artifactId>kyuubi-ha</artifactId>
             <version>${project.version}</version>
         </dependency>
