+++ kyuubi-main/src/test/scala/org/apache/kyuubi/config/AllKyuubiConfiguration.scala
@@ -26,6 +26,7 @@ import scala.collection.mutable.ArrayBuffer
 import org.apache.kyuubi.KyuubiFunSuite
 import org.apache.kyuubi.ha.HighAvailabilityConf
 import org.apache.kyuubi.metrics.MetricsConf
+import org.apache.kyuubi.zookeeper.ZookeeperConf
 
 // scalastyle:off line.size.limit
 /**
@@ -71,6 +72,7 @@ class AllKyuubiConfiguration extends KyuubiFunSuite {
     KyuubiConf
     HighAvailabilityConf
     MetricsConf
+    ZookeeperConf
 
     val newOutput = new ArrayBuffer[String]()
     newOutput += "<!-- DO NOT MODIFY THIS FILE DIRECTLY, IT IS AUTO GENERATED BY" +
