+++ kyuubi-zookeeper/src/test/scala/org/apache/kyuubi/zookeeper/EmbeddedZookeeperSuite.scala
@@ -33,7 +33,7 @@ class EmbeddedZookeeperSuite extends KyuubiFunSuite {
     assert(zkServer.getName === zkServer.getClass.getSimpleName)
     assert(zkServer.getServiceState === LATENT)
     val conf = KyuubiConf()
-    conf.set(KyuubiConf.EMBEDDED_ZK_PORT, 0)
+    conf.set(ZookeeperConf.ZK_CLIENT_PORT, 0)
     zkServer.stop() // only for test coverage
     zkServer.initialize(conf)
     assert(zkServer.getConf === conf)
@@ -49,7 +49,7 @@ class EmbeddedZookeeperSuite extends KyuubiFunSuite {
   test("connect test with embedded zookeeper") {
     val zkServer = new EmbeddedZookeeper()
     intercept[AssertionError](zkServer.getConnectString)
-    zkServer.initialize(KyuubiConf().set(KyuubiConf.EMBEDDED_ZK_PORT, 0))
+    zkServer.initialize(KyuubiConf().set(ZookeeperConf.ZK_CLIENT_PORT, 0))
     zkServer.start()
 
     val zkClient = CuratorFrameworkFactory.builder()
